# postbox-contract/openapi.yaml
openapi: 3.1.0
info:
  title: Postamat Platform API
  version: 1.0.0-draft.1
  summary: Parcel locker platform for Türkmenpoçta — client app, kiosk terminal, admin panel.
  description: |
    Single source of truth for the HTTP contract between three front-ends
    (Flutter client app, kiosk terminal, admin panel) and the core backend.

    Conventions — see `docs/API-01.md` for prose:

    * All JSON fields are `snake_case`.
    * All identifiers are UUID (v7 preferred).
    * All timestamps are UTC, ISO-8601 with `Z`. Clients render in `Asia/Ashgabat`.
    * All money is an integer in minor units plus a currency code. Never a float.
    * All phone numbers are E.164 (`+993…`).
    * Breaking changes ship as `/api/v2`, never as edits to `/api/v1`.

    Generated clients: `orval`/`openapi-typescript` for web, `openapi-generator` for Dart.
    Do not hand-edit generated code — edit this file.
  contact:
    name: Postamat platform team
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: https://api.postamat.tm/api/v1
    description: Production (inside Türkmenistan)
  - url: https://stage.postamat.tm/api/v1
    description: Staging
  - url: http://localhost:4010/api/v1
    description: Prism mock from this file
...
security:
  - clientAuth: []
paths:
...
  /bookings:
    get:
      tags: [Bookings]
      operationId: listBookings
      summary: The client's bookings
      description: |
        Backs the two tabs on screen `488:8413` / `270:5270`.
        `scope=active` returns everything not yet finished; `scope=history` the rest.
      parameters:
        - name: scope
          in: query
          schema:
            type: string
            enum: [active, history]
            default: active
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Bookings, newest first.
          content:
            application/json:
              schema:
                type: object
                required: [items, pagination]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/BookingListItem'
                  pagination:
                    $ref: '#/components/schemas/CursorPagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Bookings]
      operationId: createBooking
      summary: Create a booking and hold a cell
      description: |
        A concrete cell is picked and held immediately — the client is about to pay for
        it, so it cannot be handed to somebody else in the meantime. The hold lasts
        `hold_expires_at - now` (10 minutes by default) and the cell sits in `booked`.
        If payment does not complete in time the hold is released and the cell returns
        to `free`.

        `recipient_phone` may be omitted — that is the «Пропустить» path on screen
        `262:3414`, meaning the sender will collect the parcel themselves.

        The sender always pays, always before the parcel goes in. There is no
        pay-on-collection flow.

        There is no fixed cap on concurrent bookings, but the limit is a server
        setting rather than a hard-coded absence: when it is switched on, this endpoint
        answers `BOOKING_LIMIT_EXCEEDED` and the app renders that message. Clients
        must handle the code from day one even while the limit is disabled.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [postamat_id, cell_type_id, duration_hours]
              properties:
                postamat_id:
                  type: string
                  format: uuid
                cell_type_id:
                  type: string
                  format: uuid
                duration_hours:
                  $ref: '#/components/schemas/DurationHours'
                recipient_phone:
                  oneOf:
                    - $ref: '#/components/schemas/Phone'
                    - type: 'null'
                recipient_name:
                  type: [string, 'null']
                  maxLength: 100
            example:
              postamat_id: 0190f2c1-6b7a-7c3e-9a11-2f0c9a5d4e20
              cell_type_id: 0190f2c1-6b7a-7c3e-9a11-2f0c9a5d4e30
              duration_hours: 24
              recipient_phone: '+99362123456'
      responses:
        '201':
          description: |
            Created and held. The response carries the assigned cell number so the
            summary screen (`262:3609`) can show it before payment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: |
            Codes: `NO_FREE_CELLS`, `CELL_TYPE_UNAVAILABLE`, `POSTAMAT_BLOCKED`.
            `NO_FREE_CELLS` carries nearby alternatives so the app can offer them.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: NO_FREE_CELLS
                  message: No cells of this size are free.
                  details:
                    alternatives:
                      - postamat_id: 0190f2c1-6b7a-7c3e-9a11-2f0c9a5d4e21
                        name: ТП #3 Багтыярлык
                        distance_m: 2200
                        available_count: 2
                  trace_id: 0190f2c1-6b7a-7c3e-9a11-2f0c9a5d4e40
        '422':
          description: |
            Codes: `DURATION_NOT_SUPPORTED`, `RECIPIENT_PHONE_INVALID`,
            `RECIPIENT_PHONE_SAME_AS_SENDER`, `PROFILE_INCOMPLETE`,
            `BOOKING_LIMIT_EXCEEDED`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bookings/{booking_id}:
    get:
      tags: [Bookings]
      operationId: getBooking
      summary: One booking
      description: |
        Everything screen `262:3977` needs, including `deposit_code` while it is still
        valid. The pickup code is never present in this payload.
      parameters:
        - $ref: '#/components/parameters/BookingId'
      responses:
        '200':
          description: Booking.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /bookings/{booking_id}/cancel:
    post:
      tags: [Bookings]
      operationId: cancelBooking
      summary: Cancel a booking
      description: |
        Allowed while the booking is unpaid, or paid but the parcel has not been
        deposited yet. Once a courier has closed the door the cell is occupied and
        cancellation is no longer a client-side action.

        **Cancelling a paid booking does not refund anything in this version.**
        The cell is released, the money stays. There is no refund endpoint anywhere in
        this API by design — that has to be spelled out on the summary screen before
        the client pays, not discovered afterwards.
      parameters:
        - $ref: '#/components/parameters/BookingId'
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: [string, 'null']
                  maxLength: 500
      responses:
        '200':
          description: Cancelled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: |
            Codes: `BOOKING_NOT_CANCELLABLE`, `BOOKING_ALREADY_CANCELLED`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bookings/{booking_id}/extend-hold:
    post:
      tags: [Bookings]
      operationId: extendBookingHold
      summary: Extend the 10-minute payment hold
      description: |
        For the case where the hold runs out while the client is on the bank's 3-D
        Secure page. Limited number of extensions; after that the hold dies for good.
      parameters:
        - $ref: '#/components/parameters/BookingId'
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      responses:
        '200':
          description: Hold extended.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '409':
          description: |
            Codes: `BOOKING_HOLD_EXPIRED`, `HOLD_EXTENSION_LIMIT_EXCEEDED`,
            `BOOKING_ALREADY_PAID`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bookings/{booking_id}/timeline:
    get:
      tags: [Bookings]
      operationId: getBookingTimeline
      summary: Booking timeline
      description: |
        The checklist on screen `262:3977`: booked → parcel deposited → pickup code
        sent → collected. Steps that have not happened yet are returned with
        `occurred_at: null` so the UI can render them greyed out in the right order.
      parameters:
        - $ref: '#/components/parameters/BookingId'
      responses:
        '200':
          description: Timeline.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/TimelineStep'
              example:
                items:
                  - step: booked
                    occurred_at: '2026-08-14T06:04:00Z'
                  - step: parcel_deposited
                    occurred_at: '2026-08-14T06:04:00Z'
                  - step: pickup_code_sent
                    occurred_at: null
                  - step: collected
                    occurred_at: null
        '404':
          $ref: '#/components/responses/NotFound'

  # ────────────────────────────────────────────────────────────────── Payments
...
  /bookings/{booking_id}/deposit-code/rotate:
    post:
      tags: [Access codes]
      operationId: rotateDepositCode
      summary: Issue a fresh deposit code
      description: |
        Button «Повторно отправить PIN-код курьера». There is no "re-send the same
        code" operation anywhere in this API, because the server stores a keyed hash
        and cannot reproduce the digits. Asking again therefore **mints a new code and
        kills the old one** — which is also the safer behaviour: a code that was
        forwarded to the wrong person stops working the moment the sender asks for
        another.

        The new code comes back in the response (so the app can show and copy it) and
        goes out by SMS to `courier_phone`.

        The pickup code has no equivalent read: it is generated server-side and
        delivered by SMS to the recipient only after the door sensor confirms the
        parcel is inside.
      parameters:
        - $ref: '#/components/parameters/BookingId'
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  oneOf:
                    - $ref: '#/components/schemas/Phone'
                    - type: 'null'
                  description: |
                    Send to a different courier and store that number. Omit to reuse
                    the one captured on screen `262:3414`.
      responses:
        '200':
          description: New code issued; the previous one is revoked.
          content:
            application/json:
              schema:
                type: object
                required: [code, expires_at, resend_after]
                properties:
                  code:
                    $ref: '#/components/schemas/Pin'
                  expires_at:
                    $ref: '#/components/schemas/Timestamp'
                  resend_after:
                    type: integer
                    description: Seconds before another rotation is allowed.
                    example: 60
              example:
                code: '45865'
                expires_at: '2026-08-15T09:41:00Z'
                resend_after: 60
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: |
            Codes: `GRANT_ALREADY_USED` — the parcel is already inside, so there is
            nothing left to open with; `GRANT_EXPIRED`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: 'Codes: `CODE_RESEND_TOO_SOON`, `CODE_RESEND_LIMIT_EXCEEDED`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bookings/{booking_id}/pickup-code/rotate:
    post:
      tags: [Access codes]
      operationId: rotatePickupCode
      summary: Issue a fresh pickup code, SMS it, and return it once
      description: |
        Button «Повторно отправить одноразовый PIN-код». Same rule as the deposit code:
        the old one dies, a new one is minted, and the digits come back **once** — in
        this response and nowhere else.

        The sender reads them. That is the C2C flow the mockups draw: the recipient
        never registers anywhere, and when the SMS does not arrive the sender is the
        one who reads the code out over the phone. Returning nothing here would break
        precisely the case this endpoint exists for.

        What the two grants protect is unchanged: the **courier** holds the deposit
        code and only that, so the person carrying the box still cannot collect it. A
        sender who opens their own parcel is not theft — they paid for it.

        Every later read returns null. Never in a notification, never in the admin
        panel, never in a log line.
      parameters:
        - $ref: '#/components/parameters/BookingId'
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      responses:
        '200':
          description: |
            New code issued, the previous one revoked, SMS on its way. The digits are
            here and in no other response.
          content:
            application/json:
              schema:
                type: object
                required: [code, expires_at, resend_after]
                properties:
                  code:
                    $ref: '#/components/schemas/Pin'
                  expires_at:
                    $ref: '#/components/schemas/Timestamp'
                  resend_after:
                    type: integer
        '409':
          description: |
            Codes: `PICKUP_CODE_NOT_ISSUED_YET` (the parcel is not in the cell yet),
            `GRANT_ALREADY_USED`, `GRANT_EXPIRED`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: 'Codes: `CODE_RESEND_TOO_SOON`, `CODE_RESEND_LIMIT_EXCEEDED`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bookings/{booking_id}/pickup-code/transfer:
    post:
      tags: [Access codes]
      operationId: transferPickup
      summary: Transfer the right to collect to another phone
      description: |
        Revokes the current pickup grant and issues a fresh one to a different number.
        People forward codes over messengers anyway — this makes it an explicit,
        audited action instead of an invisible one.
      parameters:
        - $ref: '#/components/parameters/BookingId'
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [new_phone]
              properties:
                new_phone:
                  $ref: '#/components/schemas/Phone'
                new_name:
                  type: [string, 'null']
                  maxLength: 100
      responses:
        '200':
          description: |
            Transferred. The old code stops working immediately.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '409':
          description: |
            Codes: `TRANSFER_NOT_ALLOWED` (parcel already collected),
            `PICKUP_CODE_NOT_ISSUED_YET`, `GRANT_ALREADY_USED`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 'Codes: `TRANSFER_PHONE_SAME`, `RECIPIENT_PHONE_INVALID`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ───────────────────────────────────────────────────────────────────── Kiosk
...
components:
  securitySchemes:
    clientAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Client access token, 15 minutes. Renew with `/auth/refresh`.
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin access token, 15 minutes. Renew with `/admin/auth/refresh`.
    deviceAuth:
      type: http
      scheme: bearer
      description: |
        Long-lived device token obtained at provisioning. Kept in the device keystore.
    staffAuth:
      type: http
      scheme: bearer
      description: Staff-mode token on a terminal. Short-lived, dies with inactivity.
  parameters:
...
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor from the previous response. Feeds and logs only.
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
  responses:
...
    Unauthorized:
      description: |
        No or expired credentials. Codes: `UNAUTHORIZED`, `TOKEN_EXPIRED`,
        `TOKEN_INVALID`. **401 means refresh the token — never confuse it with 403.**
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    BookingListItem:
      type: object
      required: [id, status, postamat_name, cell_number, cell_type_name, created_at]
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/BookingStatus'
        postamat_id:
          type: string
          format: uuid
        postamat_name:
          type: string
          example: ТП #4 Беркарарлык
        cell_number:
          type: string
          example: '14'
        cell_type_name:
          type: string
          example: Маленький
        created_at:
          $ref: '#/components/schemas/Timestamp'
        expires_at:
          type: [string, 'null']
          format: date-time
...